added reference id to get and store external collateral reference upon loan creation

+++ contracts/strategies/base/BaseBorrowStrategy.sol
@@ -64,7 +64,7 @@ abstract contract BaseBorrowStrategy is BaseLongStrategy {
 
         // Update loan's total liquidity debt and principal amounts
         uint256 loanLiquidity = _loan.liquidity;
-        _loan.px = updateLoanPrice(liquidityBorrowed, getCurrentCFMMPrice(), loanLiquidity, _loan.px);
+        _loan.px = uint248(updateLoanPrice(liquidityBorrowed, getCurrentCFMMPrice(), loanLiquidity, _loan.px));
         liquidity = loanLiquidity + liquidityBorrowed;
         _loan.initLiquidity = _loan.initLiquidity + uint128(liquidityBorrowed);
         _loan.lpTokens = _loan.lpTokens + lpTokens;
